requirements.txt
asgiref==3.7.2
Brotli==1.0.9
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cssbeautifier==1.14.8
dj-database-url==2.0.0
Django==4.2.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-rename-app==0.1.6
django-taggit==4.0.0
django-tinymce==3.6.1
djlint==1.31.0
EditorConfig==0.12.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.8
json5==0.9.14
pathspec==0.11.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
whitenoise==6.5.0
